# Fabric/Fabrics.WebApplication/API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fabrics.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Query/shipments-by-provider": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Get all shipments by a specific provider.",
        "parameters": [
          {
            "name": "providerId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShipmentGetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Query/providers-by-date": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Get all providers who supplied goods within a specific period, ordered by name.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderGetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Query/max-quantity-providers-by-period": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Get providers with the maximum quantity of goods supplied within a specified period.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProviderGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id is used to store the ID.",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name is used to store name of Provider.",
            "nullable": true
          },
          "typeOfGoods": {
            "type": "string",
            "description": "TypeOfGoods is used to store product type information.",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Address is used to store address of Provider.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Class ProviderGetDto is used to make HTTP GET request."
      },
...
      "ShipmentGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id is used to store the ID.",
            "format": "int32"
          },
          "fabricId": {
            "type": "integer",
            "description": "FabricId is used to store the ID of recipient.",
            "format": "int32"
          },
          "providerId": {
            "type": "integer",
            "description": "ProviderId is used to store the ID of sender.",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "description": "Date is used to store the information about date of shipment.",
            "format": "date-time"
          },
          "numberOfGoods": {
            "type": "integer",
            "description": "NumberOfGoods is used to store the information about amount of goods in shipment.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Class ShipmentGetDto is used to make HTTP GET request."
      },
...
    }
  }
}